% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_word_count.R
\name{book_word_count}
\alias{book_word_count}
\title{Book Word Count}
\usage{
book_word_count(update = FALSE)
}
\arguments{
\item{update}{Boolean, FALSE by default. If TRUE, also takes the sum of the
word counts without code-chunks and updates the field \code{book_word_count}
in \verb{_meta.yml}.}
}
\value{
A dataframe of word and characters counts from all book input files.
}
\description{
Returns a dataframe with the word counts for index.Rmd and all input files
in the book input directory, along with totals, using rmdwc::rmdcount and
janitor::adorn_totals.
}
\examples{
book_word_count()
}
