% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_book.R
\name{build_book}
\alias{build_book}
\title{Build book}
\usage{
build_book(format = "print", draft = TRUE)
}
\arguments{
\item{format}{String, default "print". One of "all", "print", "gitbook", "pdf",
or "word". "all" produces gitbook, pdf, and word outputs. "print" produces just
pdf and word outputs. Individual output selections produces just that output.}

\item{draft}{Boolean, default TRUE. If TRUE, build the book in the 'book_draft/DATE'
folder, where DATE is today's date. Set to FALSE to build the final book in
'book_output', ready for use.}
}
\description{
Build book
}
\examples{

}
